ci: switch to Github Actions (#18)
This converts the repo to Github Actions for CI.

Of note, the Github windows runner does not support containers, so it's
using a Chocolatey package for Redis.

Additionally, to get the redis cluster networking to work properly when
using the `services` config, I had to run the CI job in a container
rather than on the `ubuntu-latest` runner itself. So, it's running in
`ubuntu:22.04` with essential packages installed.