ci: fix windows tests (#20)
In my original PR that introduced Github Actions, the tests on windows
were not actually succeeding. This issue was masked by `go test` caching
previous (successful) results.

This PR fixes the tests by running `redis-server` in the background, and
additionally adds a `count=1` argument to `go tests` to ensure this type
of issue does not happen again.